requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1

# ── Socket.IO / async transport ───────────────────────────────────────────────
python-engineio==4.11.2
python-socketio==5.12.1
eventlet==0.39.1

# ── WSGI server (production) ──────────────────────────────────────────────────
gunicorn==23.0.0

# ── Rate limiting ─────────────────────────────────────────────────────────────
Flask-Limiter==3.9.0

# ── Input validation ──────────────────────────────────────────────────────────
jsonschema==4.23.0

# ── Scientific stack ──────────────────────────────────────────────────────────
numpy>=1.26.4
scipy>=1.13.1
pandas>=2.2.2
matplotlib>=3.8.4

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
Werkzeug>=3.1.3
psutil>=6.0.0

# ── Development / test extras (install separately or via extras_require) ──────
# pytest>=8.0
# pytest-flask>=1.3
# pytest-cov>=5.0
